Remove unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/jakarta/jmeter/trunk@325350 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: c01ff2884f6409cb2c0be07f1cddc7ca301111ce

+++ src/core/org/apache/jmeter/testelement/property/LongProperty.java
@@ -111,7 +111,7 @@ public class LongProperty extends NumberProperty
      */
     public double getDoubleValue()
     {
-        return (double) value;
+        return value;
     }
 
     /**
@@ -119,7 +119,7 @@ public class LongProperty extends NumberProperty
      */
     public float getFloatValue()
     {
-        return (float) value;
+        return value;
     }
 
     /**
